refactor(viewer): give the webgl_unavailable family its own module
`load-errors.ts` had grown to 456 lines, past the ~400-line house rule, and
the new three.js refusal matcher is what pushed it over (Codex review, #2459).

Two extractions, both wordings-only — no matcher's behaviour, anchoring or
call order changes, and no test moves:

- `lib/webgl-unavailable.ts` now holds the WHOLE `webgl_unavailable` family:
  three's refusal wordings (moved verbatim, rationale included) next to the
  import of MapLibre's, behind one `isWebglUnavailable`. Previously the two
  arms were in different files and only met in an `||` inside
  `classifyLoadError`, which is how an ungrouped third matcher gets written;
  now the set is one read, and the module header says an added arm must be
  anchored.
- `lib/cancelled-and-network-errors.ts` takes `cancelled` and
  `network_unavailable`, which share `STRINGIFIED_ERROR_PREFIX` (used by
  nothing else), already document themselves by contrast with each other, and
  carry the same #2410 anchoring invariant.

`load-errors.ts` keeps the taxonomy and the order the buckets are tried in,
and drops to 294 lines; the new modules are 92 and 143.
