+++ apps/viewer/src/lib/load-errors.ts
@@ -3,12 +3,17 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 /**
- * Classification of model-load failures. The user-facing humanisation lives in
- * ./load-error-message.ts.
+ * Classification of model-load failures. This module owns the taxonomy and the
+ * ORDER the buckets are tried in; the user-facing humanisation lives in
+ * ./load-error-message.ts, and the two families whose matchers carry more
+ * rationale than pattern have their own modules (./webgl-unavailable.ts,
+ * ./cancelled-and-network-errors.ts) rather than a longer file here.
  *
  * Despite the name, this is the viewer's ONLY error-family classifier —
  * `analytics-scrub.ts` runs it over every captured `$exception` — so a non-load
- * family needing grouping belongs here too, not in a second one that would drift.
+ * family needing grouping belongs here too, not in a second one that would
+ * drift. The sibling modules are not that: they hold a family's WORDINGS, while
+ * every kind and the order it is tried in stays in {@link classifyLoadError}.
  *
  * The geometry/parser workers both initialise the same `@ifc-lite/wasm` binary.
  * wasm-bindgen's streaming loader rethrows on a non-OK HTTP status (it only
@@ -19,9 +24,8 @@
  * ./load-error-message.ts turns that kind into the user-facing message.
  */
 
-// Imported, not restated: a second copy of MapLibre's wordings would drift on
-// the next upgrade. That module imports nothing, so this adds no dependency.
-import { isWebglContextCreationError } from './geo/map-webgl-support.js';
+import { isCancelledError, isNetworkUnavailableError } from './cancelled-and-network-errors.js';
+import { isWebglUnavailable } from './webgl-unavailable.js';
 
 /** Stable, analytics-friendly classification of a load failure. */
 export type LoadErrorKind =
@@ -68,7 +72,8 @@ export type LoadErrorKind =
   | 'cancelled'
   /**
    * A fetch failed at the transport layer and the browser told us nothing else
-   * (the per-engine wordings are enumerated on `BARE_TRANSPORT_FAILURE` below).
+   * (the per-engine wordings are enumerated on `BARE_TRANSPORT_FAILURE`, in
+   * ./cancelled-and-network-errors.ts).
    * The connection dropped, went offline, or was killed mid-flight. Nothing in
    * the app is broken, so this is deliberately the LAST bucket checked: any
    * failure that identified itself keeps its own kind.
@@ -79,12 +84,9 @@ export type LoadErrorKind =
    * either `/mcp` three.js scene (#2458). Not a load failure and never reaches
    * `formatLoadError`; classified so the family gets ONE fingerprint instead of
    * one issue per deploy and per wording. Membership is by MESSAGE, not by who
-   * caught it: `isWebglContextCreationError` matches only MapLibre's own
-   * wordings and the two strings `LocationMap` synthesizes, and
-   * `isThreeContextRefusal` only three.js's two authored wordings and the one
-   * `three-webgl-support.ts` synthesizes — all of them anchored, so an error
-   * that merely MENTIONS the phrase keeps its own identity and its `error`
-   * severity.
+   * caught it, and both libraries' wordings live together in
+   * ./webgl-unavailable.ts — anchored, so an error that merely MENTIONS one of
+   * the phrases keeps its own identity and its `error` severity.
    */
   | 'webgl_unavailable'
   /** Anything else. */
@@ -221,168 +223,6 @@ function isWasmRuntimeCrashError(err: unknown, message: string): boolean {
   );
 }
 
-/**
- * A cross-realm throwable reaches the analytics path as `String(err)` — i.e.
- * `"TypeError: Load failed"`, not `"Load failed"`. Structural, so the anchored
- * matchers below tolerate it; BOUNDED, so it cannot stand in for an arbitrary
- * leading sentence of ours that happens to end in "…Error:".
- *
- * `{0,32}`, not `{1,32}`: the commonest constructor is `Error` itself, whose
- * stringification is the bare `Error: Load failed`, and requiring a letter in
- * front excluded exactly that (Codex review, #2431). What does the real work is
- * `[A-Za-z]` admitting no space and no colon — a leading clause of ours can
- * never be swallowed however the count is written.
- */
-const STRINGIFIED_ERROR_PREFIX = '(?:[A-Za-z]{0,32}Error:\\s*)?';
-
-/**
- * The user (or a superseding load) cancelled the operation.
- *
- * ANCHORED, because `cancelled` is in `BENIGN_ERROR_KINDS` (./analytics-scrub.ts):
- * matching it downgrades the event to `warning` and fingerprints it into the
- * cancellation issue, taking an actionable failure off the error-level list as
- * effectively as deleting it. The old `/\bcancel(?:led|ed)?\b/` fired on the word
- * ANYWHERE — and, both suffixes optional, on a bare "cancel" — so `Upload failed:
- * driver shim logged cancelled while retrying` read benign (#2410).
- *
- * Shaped differently from the whole-message anchor `isNetworkUnavailableError`
- * uses, because the wordings have a different author: the transport strings
- * come from `fetch()` and ARE the entire message, whereas cancellations are
- * mostly ours and carry authored detail after the token (`Sync cancelled:
- * <model> was removed while its update was downloading.`, `Clash run cancelled
- * before meshing.`). So what is pinned is the SUBJECT — a cancellation report
- * names what was cancelled in its opening words; a failure that merely mentions
- * one reaches the token past a clause of its own. Two arms, because the bare
- * token has no subject to name and so gets no trailing latitude at all
- * (otherwise `cancelled and our upload pipeline then wrote 0 bytes` walks
- * through). The residual, stated plainly: `Upload cancelled and the pipeline
- * wrote 0 bytes` is still read as a cancellation report, because it is one.
- */
-const CANCELLED_REPORT = new RegExp(
-  `^\\s*${STRINGIFIED_ERROR_PREFIX}(?:`
-  + 'cancell?ed\\.?\\s*$'
-  // `[^\s:]+` so the subject cannot be reached across a `:` — that separator is
-  // what makes "Upload failed: …" a sentence about the upload, not a cancellation.
-  + '|(?:[^\\s:]+\\s+){1,2}cancell?ed\\b'
-  + ')',
-  'i',
-);
-
-/**
- * The browsers' own abort wordings — unlike ours these ARE the whole message,
- * so both ends are pinned. `.name === 'AbortError'` already claims the live
- * DOMException in {@link classifyLoadError}; this covers the analytics path,
- * where all we ever have is the stringified value.
- */
-const ABORTED_MESSAGE = new RegExp(
-  `^\\s*${STRINGIFIED_ERROR_PREFIX}(?:`
-  + 'the operation was aborted'          // Gecko
-  + '|the user aborted a request'        // Chromium
-  + '|fetch is aborted'                  // WebKit
-  + '|signal is aborted without reason'  // AbortController with no reason
-  + ')\\.?\\s*$',
-  'i',
-);
-
-function isCancelledError(message: string): boolean {
-  // `AbortError` is the stringified NAME — an identity, not prose — so it is
-  // anchored to the start rather than matched as a substring anywhere.
-  return CANCELLED_REPORT.test(message)
-    || /^\s*AbortError\b/.test(message)
-    || ABORTED_MESSAGE.test(message);
-}
-
-/**
- * A bare transport failure: the request never completed and the browser gave us
- * nothing but its house phrasing. These strings originate inside `fetch()`
- * rather than in our frames, so they arrive with an EMPTY stack — exactly how
- * #1903 reached error tracking as an unattributable `TypeError: Load failed`.
- * Checked LAST, so a failure that named itself (the engine binary, the file, a
- * worker) keeps its own, more actionable kind.
- *
- * ANCHORED AT BOTH ENDS, and that is load-bearing rather than tidiness: this is
- * the most dangerous label in the file. `analytics-scrub.ts` downgrades it to
- * `warning` via `BENIGN_ERROR_KINDS` AND deletes the event outright when the
- * browser also reported `navigator.onLine === false`, so the old substring test
- * meant any failure of ours quoting a transport phrase (`Upload failed: driver
- * shim logged Failed to fetch while retrying`) was silenced, and on an offline
- * client destroyed with no record (#2410). The anchor is exactly the premise
- * stated above — these strings come FROM `fetch()`, so anything wrapping one is
- * by construction ours. A wrapped transport failure now falls to `unknown`:
- * loud, own grouping, the safe direction for a droppable kind.
- *
- * This subsumes what used to be an explicit module-import exclusion: Chromium's
- * `Failed to fetch dynamically imported module: …/assets/Foo-<hash>.js` is our
- * deploy rotating an asset under a still-open tab, not a dropped connection, and
- * it names the module, so it is no longer the whole wording. The guard that said
- * so is gone rather than left as an unreachable branch — `load-errors.test.ts`
- * keeps the regression test as the live gate on this anchor.
- */
-const BARE_TRANSPORT_FAILURE = new RegExp(
-  `^\\s*${STRINGIFIED_ERROR_PREFIX}(?:`
-  // WebKit/Safari, Chromium, Gecko — the generic "fetch rejected" strings.
-  + 'load failed'
-  + '|failed to fetch'
-  // Gecko's full wording is `NetworkError when attempting to fetch resource.`;
-  // the noun is optional only because the old matcher keyed on the fragment.
-  + '|networkerror when attempting to fetch(?:\\s+resource)?'
-  // Darwin's CFNetwork wordings, surfaced verbatim by Safari/WebKit when the
-  // connection drops, the device is offline, or DNS cannot resolve the host.
-  + '|the network connection was lost'
-  + '|the internet connection appears to be offline'
-  + '|a server with the specified hostname could not be found'
-  + ')\\.?\\s*$',
-  'i',
-);
-
-function isNetworkUnavailableError(message: string): boolean {
-  return BARE_TRANSPORT_FAILURE.test(message);
-}
-
-/**
- * The same device refusal, reached through three.js instead of MapLibre.
- *
- * `WebGLRenderer`'s constructor asks the canvas for `webgl2` and throws one of
- * exactly two authored messages when the browser says no (three@0.185.1):
- *
- *   'THREE.WebGLRenderer: Error creating WebGL context.'
- *   'THREE.WebGLRenderer: Error creating WebGL context with your selected attributes.'
- *
- * plus the third that `components/mcp/three-webgl-support.ts` synthesises when
- * its pre-flight probe answers before three is ever constructed — the arm that
- * fires FIRST on a device with no `webgl2`, and therefore the one production
- * mostly sees.
- *
- * Claimed here only because #2401 changed what these mean. #2354 deliberately
- * left them out: back then they escaped a mount effect and took the whole
- * `/mcp` route down through `ChunkErrorBoundary`, and folding a page-killing
- * crash into the benign minimap family would have hidden it. Both `/mcp` scenes
- * now mount behind `useThreeScene`, the only `WebGLRenderer` construction sites
- * in this app, so these strings can only arrive as the handled, one-per-session
- * degradation the minimap already files. Same device condition, same
- * fingerprint, and the same `warning` severity — `webgl_unavailable` is in
- * `BENIGN_ERROR_KINDS`, so this also stops one dead panel competing with real
- * breakage on the error-level list. `context` / `three_surface` still separate
- * the two surfaces inside the issue.
- *
- * ANCHORED, like every matcher here, and STRICTER than
- * `isThreeWebglContextError` in `three-webgl-support.ts`, which prefix-matches:
- * that one decides whether to degrade one panel and is recoverable when it
- * over-matches; this one hands out a shared fingerprint and is not.
- */
-const THREE_CONTEXT_REFUSAL = new RegExp(
-  '^\\s*THREE\\.WebGLRenderer: Error creating WebGL context'
-  + '(?: with your selected attributes)?\\.'
-  // The pre-flight suffix, kept in step with `threeProbeFailureError()`. A
-  // drift canary in three-webgl-support.test.ts classifies that function's
-  // real output, so a reworded probe cannot silently fall out of the family.
-  + '(?: \\(pre-flight probe\\))?\\s*$',
-);
-
-function isThreeContextRefusal(message: string): boolean {
-  return THREE_CONTEXT_REFUSAL.test(message);
-}
-
 /** Classify a load failure into a stable analytics bucket. */
 export function classifyLoadError(err: unknown): LoadErrorKind {
   const message = messageOf(err);
@@ -403,11 +243,9 @@ export function classifyLoadError(err: unknown): LoadErrorKind {
   if (name === 'AbortError') return 'cancelled';
   // BEFORE the memory/network buckets: MapLibre's failure carries the driver's
   // own `statusMessage`, vendor prose we do not control and free to contain the
-  // words those matchers key on ("allocation failed"). Safe to claim first — it
-  // takes only MapLibre's authored wordings and its event token, not a name.
-  if (isWebglContextCreationError(err) || isThreeContextRefusal(message)) {
-    return 'webgl_unavailable';
-  }
+  // words those matchers key on ("allocation failed"). Safe to claim first —
+  // every arm of it is anchored or structural (see ./webgl-unavailable.ts).
+  if (isWebglUnavailable(err, message)) return 'webgl_unavailable';
   if (isWasmEngineLoadError(message)) return 'wasm_engine_load';
   // Explicit memory-exhaustion signals win over the worker-crash bucket so a
   // worker that died with a clear OOM message is grouped as out_of_memory.
